+++ scripts/processors/4-word.sh
@@ -1,34 +1,29 @@
 #!/bin/bash
 # Author: crypt0rr - https://github.com/crypt0rr/
 
-# CTRL-C catch + cleanup of temp files
-function clean_up {
-    rm $tmp 2>/dev/null
-    source hash-cracker.sh
-}
-
-trap clean_up SIGINT SIGTERM
-
 # Requirements
-if [[ "$STATICCONFIG" = true ]]; then
-    source hash-cracker.conf
-else
-    source scripts/selectors/hashtype.sh
-    source scripts/selectors/hashlist.sh
-fi
+processor_bootstrap
 
 # Rules
 source scripts/rules/rules.config
-RULELIST=($tenKrules $NSAKEYv2 $fordyv1 $pantag $OUTD $techtrip2 $williamsuper $digits3 $dive)
+RULELIST=("$tenKrules" "$NSAKEYv2" "$fordyv1" "$pantag" "$OUTD" "$techtrip2" "$williamsuper" "$digits3" "$dive")
 
 # Temporary Files
-tmp=$(mktemp /tmp/hash-cracker-tmp.XXXX)
+tmp=$(dryrun_tempfile word)
+trap 'processor_interrupt "$tmp"' INT TERM
+trap 'processor_cleanup "$tmp"' EXIT
 
 # Logic
 read -p "Enter a word (e.g. company name): " WORD
-echo $WORD > $tmp
-for RULE in ${RULELIST[*]}; do
-    $HASHCAT $KERNEL --bitmap-max=24 -d $DEVICE $HWMON $SHOWCRACKED --potfile-path=$POTFILE -m$HASHTYPE $HASHLIST $tmp -r $RULE $LOOPBACK
+if dry_run_enabled; then
+    dryrun_note "would write custom word input to $tmp"
+else
+    echo $WORD >$tmp
+fi
+for RULE in "${RULELIST[@]}"; do
+    hashcat_base $tmp -r $RULE $LOOPBACK
 done
-rm $tmp
+if ! dry_run_enabled; then
+    rm $tmp
+fi
 echo -e "\nWord processing done\n"
