+++ scripts/processors/5-word-bruteforce.sh
@@ -1,31 +1,28 @@
 #!/bin/bash
 # Author: crypt0rr - https://github.com/crypt0rr/
 
-# CTRL-C catch + cleanup of temp files
-function clean_up {
-    source hash-cracker.sh
-    rm $tmp 2>/dev/null
-}
-
 # Requirements
-if [[ "$STATICCONFIG" = true ]]; then
-    source hash-cracker.conf
-else
-    source scripts/selectors/hashtype.sh
-    source scripts/selectors/hashlist.sh
-fi
+processor_bootstrap
 
 # Temporary Files
-tmp=$(mktemp /tmp/hash-cracker-tmp.XXXX)
+tmp=$(dryrun_tempfile word)
+trap 'processor_interrupt "$tmp"' INT TERM
+trap 'processor_cleanup "$tmp"' EXIT
 
 # Logic
 read -p "Enter a word (e.g. company name): " WORD
-echo $WORD > $tmp
-$HASHCAT $KERNEL --bitmap-max=24 -d $DEVICE $HWMON $SHOWCRACKED --potfile-path=$POTFILE -m$HASHTYPE $HASHLIST -a6 $tmp '?d?d?d?d?d?d?d?d' -i
-$HASHCAT $KERNEL --bitmap-max=24 -d $DEVICE $HWMON $SHOWCRACKED --potfile-path=$POTFILE -m$HASHTYPE $HASHLIST -a6 $tmp '?l?l?l?l?l?l' -i
-$HASHCAT $KERNEL --bitmap-max=24 -d $DEVICE $HWMON $SHOWCRACKED --potfile-path=$POTFILE -m$HASHTYPE $HASHLIST -a7 '?d?d?d?d?d?d?d?d' $tmp -i
-$HASHCAT $KERNEL --bitmap-max=24 -d $DEVICE $HWMON $SHOWCRACKED --potfile-path=$POTFILE -m$HASHTYPE $HASHLIST -a7 '?l?l?l?l?l?l' $tmp -i
-$HASHCAT $KERNEL --bitmap-max=24 -d $DEVICE $HWMON $SHOWCRACKED --potfile-path=$POTFILE -m$HASHTYPE $HASHLIST -a6 $tmp '?a?a?a?a?a' -i
-$HASHCAT $KERNEL --bitmap-max=24 -d $DEVICE $HWMON $SHOWCRACKED --potfile-path=$POTFILE -m$HASHTYPE $HASHLIST -a7 '?a?a?a?a?a' $tmp -i
-rm $tmp
+if dry_run_enabled; then
+    dryrun_note "would write custom word input to $tmp"
+else
+    echo $WORD >$tmp
+fi
+hashcat_base -a6 $tmp '?d?d?d?d?d?d?d?d' -i
+hashcat_base -a6 $tmp '?l?l?l?l?l?l' -i
+hashcat_base -a7 '?d?d?d?d?d?d?d?d' $tmp -i
+hashcat_base -a7 '?l?l?l?l?l?l' $tmp -i
+hashcat_base -a6 $tmp '?a?a?a?a?a' -i
+hashcat_base -a7 '?a?a?a?a?a' $tmp -i
+if ! dry_run_enabled; then
+    rm $tmp
+fi
 echo -e "\nWord processing done\n"
